+++ artiq/examples/master/idle_kernel.py
@@ -5,10 +5,10 @@ class IdleKernel(EnvExperiment):
     def build(self):
         self.setattr_device("core")
         self.setattr_device("led")
-    
+
     @kernel
     def run(self):
-        start_time = now_mu() + seconds_to_mu(500*ms)
+        start_time = now_mu() + self.core.seconds_to_mu(500*ms)
         while self.core.get_rtio_counter_mu() < start_time:
             pass
         self.core.reset()