+++ artiq/runtime.rs/libksupport/api.rs
@@ -74,7 +74,7 @@ static mut API: &'static [(&'static str, *const ())] = &[
 
     /* libm */
     api!(sqrt),
-    api!(lround),
+    api!(round),
 
     /* exceptions */
     api!(_Unwind_Resume),
@@ -104,17 +104,6 @@ static mut API: &'static [(&'static str, *const ())] = &[
     api!(rtio_output),
     api!(rtio_input_timestamp),
     api!(rtio_input_data),
-
-    #[cfg(rtio_dds_count)]
-    api!(dds_init),
-    #[cfg(rtio_dds_count)]
-    api!(dds_init_sync),
-    #[cfg(rtio_dds_count)]
-    api!(dds_batch_enter),
-    #[cfg(rtio_dds_count)]
-    api!(dds_batch_exit),
-    #[cfg(rtio_dds_count)]
-    api!(dds_set),
 
     api!(i2c_init),
     api!(i2c_start),